requirements.txt
alembic==1.8.1
cent==4.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
Werkzeug==2.2.2
